A lot of stuff
- Added an app icon
- Added display scaling option
- Added multiple selection (a little broken)
- Added some translations
- Probably more
- Done somewhere in early-mid April

+++ core/src/main/java/net/thewinnt/planimetry/util/FontProvider.java
@@ -5,5 +5,5 @@ import com.badlogic.gdx.graphics.g2d.BitmapFont;
 
 @FunctionalInterface
 public interface FontProvider {
-    public BitmapFont getFont(int size, Color color);
+    public BitmapFont getFont(float size, Color color);
 }
